requirements.txt
# Core dependencies
feedparser>=6.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0

# LLM and AI
google-genai>=1.7.0
openai>=1.0.0
anthropic>=0.5.0

# Video processing
moviepy>=1.0.0
opencv-python>=4.5.0
Pillow>=8.0.0

# YouTube API
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.0

# Telegram bot
python-telegram-bot>=13.0.0
telethon>=1.24.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
slugify>=0.0.1
schedule>=1.0.0
requests>=2.31.0
Pillow>=10.0.0

# Translation (if needed)
transformers>=4.20.0
torch>=1.9.0

# Web scraping (additional)
selenium==4.22.0
webdriver-manager>=3.5.0

# Animation (if using GSAP equivalent)
selenium>=4.0.0  # for browser automation

# Logging and monitoring
loguru>=0.6.0
